requirements.txt
jupyter-book>=0.15.1,<0.16.0
markdownify==0.11.*
./mammoth-1.6.1.2-py2.py3-none-any.whl
citeproc-py==0.6.*
pytest>=7.4.3,<7.5.0
beautifulsoup4==4.12.*
